# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/admin/users": {
      "post": {
        "operationId": "AdminUsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      },
      "get": {
        "operationId": "AdminUsersController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Regular",
                "Operator",
                "Admin"
              ]
            }
          },
          {
            "name": "isDeleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      },
      "patch": {
        "operationId": "AdminUsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      },
      "delete": {
        "operationId": "AdminUsersController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
    "/admin/users/{id}/restore": {
      "post": {
        "operationId": "AdminUsersController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_localLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_localRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocalUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/giveaway": {
      "post": {
        "operationId": "GiveawayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiveawayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "giveaway"
        ]
      },
      "get": {
        "operationId": "GiveawayController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "giveaway"
        ]
      }
    },
    "/giveaway/{id}": {
      "get": {
        "operationId": "GiveawayController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "giveaway"
        ]
      },
      "patch": {
        "operationId": "GiveawayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiveawayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "giveaway"
        ]
      },
      "post": {
        "operationId": "GiveawayController_claim",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "giveaway"
        ]
      }
    },
    "/admin/giveaway": {
      "post": {
        "operationId": "AdminGiveawayController_createForCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiveawayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / giveaway"
        ]
      },
      "get": {
        "operationId": "AdminGiveawayController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / giveaway"
        ]
      }
    },
    "/admin/giveaway/{id}": {
      "post": {
        "operationId": "AdminGiveawayController_createForSpecificUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiveawayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / giveaway"
        ]
      },
      "get": {
        "operationId": "AdminGiveawayController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / giveaway"
        ]
      },
      "patch": {
        "operationId": "AdminGiveawayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiveawayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin / giveaway"
        ]
      }
    }
  },
  "info": {
    "title": "Dino Gifter API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateAdminUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "Regular",
              "Operator",
              "Admin"
            ]
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "Regular",
              "Operator",
              "Admin"
            ]
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "CreateLocalUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "DinoDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "growthLabel": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "growthLabel"
        ]
      },
      "TrialDataDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "typing",
              "math",
              "puzzle"
            ]
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "type",
          "data"
        ]
      },
      "CreateGiveawayDto": {
        "type": "object",
        "properties": {
          "dino": {
            "$ref": "#/components/schemas/DinoDataDto"
          },
          "activeAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "trials": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrialDataDto"
            }
          },
          "isCanceled": {
            "type": "boolean"
          }
        },
        "required": [
          "dino"
        ]
      },
      "UpdateGiveawayDto": {
        "type": "object",
        "properties": {
          "dino": {
            "$ref": "#/components/schemas/DinoDataDto"
          },
          "activeAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "trials": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrialDataDto"
            }
          },
          "isCanceled": {
            "type": "boolean"
          }
        }
      }
    }
  }
}