# src/users/dto/create-local-user.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, MinLength } from "class-validator";

export class CreateLocalUserDto {
  @ApiProperty()
  @IsString()
  username: string;

  @ApiProperty({ minLength: 8 })
  @IsString()
  @MinLength(8)
  password: string;
}